apps/agent-runtime/pyproject.toml
[project]
name = "agent-runtime"
version = "0.1.0"
description = "LangGraph-based runtime for Your First Engineer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "langgraph==1.0.3",
    "langchain-core>=1.0.0,<2.0.0",
    "langgraph-checkpoint-postgres>=3.0.0,<4.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
    "langchain-anthropic>=0.1.0",
    "langchain-openai>=0.1.0",
    "structlog>=24.1.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-exporter-prometheus>=0.45b0",
    "langchain-community>=0.0.10",
    "playwright>=1.40.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0,<10.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "coverage>=7.12.0,<8.0.0",
    "mypy>=1.18.0,<1.19.0",
    "ruff>=0.8.0",
    "flake8>=7.3.0",  # Optional for legacy plugins
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_runtime"]
